fix: WaveformVisualizer dataArray 타입 캐스팅 추가

+++ components/voice-test/WaveformVisualizer.tsx
@@ -57,7 +57,7 @@ export default function WaveformVisualizer({
       // Calculate average amplitude
       let amplitude = 0;
       if (isRecording && analyser && dataArray) {
-        analyser.getByteTimeDomainData(dataArray);
+        analyser.getByteTimeDomainData(dataArray as Uint8Array);
         
         // Calculate average amplitude from time domain data
         let sum = 0;
